Add contracts and tests for Buy Me A Coffee
This change adds contracts(scripts) and tests for buying a contract owner
coffee. To buy a coffee, the user will send some ETH in form of a tip and
a message. There is a also a feature to retrieve all messages that have
been sent from users and another to withdraw all the tips that have been
received from users.